refactor: WeatherManager를 NetworkManager로 네이밍 변경 및 Networkable로 네이밍 변경
